\name{rcloud.encrypt}
\alias{rcloud.encrypt}
\alias{rcloud.decrypt}
\title{
  Encrypt/decrypt objects
}
\description{
  \code{rcloud.encrypt} encrypts an R object

  \code{rcloud.decrypt} decrypt a previously encrypted R object
}
\usage{
rcloud.encrypt(x, key = get.user.key(), cipher = "AES-256")
rcloud.decrypt(x, key = get.user.key(), cipher = attr(x, "metadata")$cipher) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{object to encrypt/decrypt}
  \item{key}{string or raw vector, key to use for encryption/decryption}
  \item{cipher}{string, cipher to use}
}
\details{
  Both functions are friendly wrappers for \code{\link{PKI.encrypt}} and
  \code{\link{PKI.decrypt}}.
  
  \code{rcloud.encrypt} takes an R object or raw vector. If it is
  an object, it is serialized using \code{\link{serialize}} before
  encryption to obtain a raw vector. Then the raw vector is encrypted.

  \code{rcloud.decrypt} takes an object previously generated by
  \code{rcloud.encrypt} and decrypts it using the provided key.

  For available ciphers see \code{\link{PKI.encrypt}}, but the most
  commonly used are \code{"AES-128"} and \code{"AES-256"}. Note that the
  key must be of sufficient length for the cipher, i.e., AES-128
  requires at least 16 bytes and AES-256 32 bytes.

  The object created by \code{rcloud.encrypt} also contains the SHA1
  checksum of the raw vector before encryption as well as the cipher, so
  that \code{rcloud.decrypt} can check the validity of the
  decryption. All metadata is stored in the \code{"metadata"} attribute
  which is compatible with storing binary assets (where supported).
}
\value{
  for \code{rcloud.encrypt} a raw vector containing the encrypted
  content and with \code{"metadata"} attribute containing at most
  elements \code{"sha1"}, \code{"encoding"} and \code{"cipher"}.
  If the object was serialized, \code{encoding="rds"}.

  for \code{rcloud.decrypt} the decrypted object
}
%\references{
%}
\author{
  Simon Urbanek
}
%\note{
%}
\seealso{
  \code{\link{PKI.encrypt}}, \code{\link{PKI.decrypt}} and \code{\link{PKI.digest}}
}
%\examples{
%}
\keyword{manip}
