\name{rcloud.session.log}
\alias{rcloud.session.log}
\title{
  Send a string to the session pane.
}
\description{
  \code{rcloud.session.log} sends a string to the RCloud session pane
}
\usage{
  rcloud.session.log(text)
}
\arguments{
  \item{text}{pre-formatted text to be echoed to the the session pane}
}
\details{
  \code{rcloud.session.log} sends text to the session pane, typically
  displayed in the lower-right corner of the RCloud IDE.
}
%\value{
%}
%\references{
%}
\author{
  Gordon Woodhull
}
%\note{
%}
%\seealso{
%}
%\examples{
%}
\keyword{manip}
